fix: fixed func of creating base coords on rad 1
